Use local media files
This commit involves pulling data from a copy I made of the Google Sheet, which has a new column with the actual filename of the file in Drive (populated with AppScript in the sheet itself... sheesh).  This then maps to a folder full of the actual files which is downloaded and extracted manually.  For now the files themselves are not committed to the git repo.