Patch for `astro-imagetools`
This patch allows specifying `fallbackFormat=false` in the `renderPicture()` and other APIs, to prevent the creation of unwanted images.  It probably breaks some uses of `astro-imagetools`, especially the use of the component interfaces, but it works for the unusual usage I'm making of this package in this project.  Will look at the possibility of raising an upstream PR, but not sure if it would actually be welcome anyway.  Also looking at the "official" Astro image component which may potentially be a better long-term bet anyway,