Big performance improvement to zoom functionality
With so many images on the page, the repeated repaints caused by ` window.getComputedStyle` in a loop were killing performance.  Calculating the transforms without repainting and then applying all at once is considerably better, and applying the transforms only to images that are currently visible is even better.  Performance of zoom when the wall is unfiltered (i.e. when all 2k+ memes are on the wall) on mobile Chrome/Safari is now tolerable.