Hacky fix to ensure OpenGraph images are built correctly
`astro-imagetools` hashes optimized images based on the filename and the display size, which means that a new image written to `assets/open-graph.jpeg` is ignored if one is already cached.  Furthermore, there is no simple way to identify and remove the cached asset in question, so this commit just clears the entire image cache before rebuilding (in the GA context this seems to cost very little anyway).

In the fullness of time this issue will likely go away if it's possible  to migrate to Astro's official `@astrojs/image` integration (although this will still require hacking the API), or, hopefully better, the (as-yet still experimental) optimized assets support.