Reinstate manual cache clearing
The growing list of patches I've made to `astro-imagetools` covers correctly hashing the built assets, but `astro-imagetools` still uses an internal cache based on only the filename and output size; this means that it won't pick up new versions of `assets/open-graph.jpeg` when they're created, as it uses old cached ones instead.  Clearing the entire cache is a quick -- if very heavy-handed -- way to address this in in the immediate term.