Refactor to generate derivative images when building the dataset
This causes vite to generate a spurious `Top-level await is not available in the configured target environment` error, and the only work-around I've found so far is to exclude `astro-imagetools` from `vite.optimizeDeps`.  This is unfortunate, as it seems to cause a large increase in start-up time.  Perhaps there's another solution.