Move srcset generation to `dataset.mjs` for RSS feed generation
Assets are generated only after routes, so without a fully-primed cache the RSS feed route won't be able to read the filesize for the enclosure length property :(  Simplest workaround I've been able to find is just to run the build twice -- once to prime the asset cache (albeit generating an empty RSS feed), and once again to ensure the RSS feed is built properly.  The cost is very small, and better solutions look like they would be considerably involved :(

These changes also mean the `pnpm print-dataset` script no longer works, but that's not a big problem.